+++ src/main/java/com/beanions/common/dao/signup/SignupMapper.java
@@ -1,5 +1,6 @@
 package com.beanions.common.dao.signup;
 
+import com.beanions.common.dto.MembersDTO;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
@@ -7,4 +8,6 @@ public interface SignupMapper {
     int checkDupId(String id);
 
     int checkDupNkname(String nkname);
+
+    int joinMember(MembersDTO member);
 }
