+++ src/main/java/com/beanions/common/service/SignupService.java
@@ -1,6 +1,7 @@
 package com.beanions.common.service;
 
 import com.beanions.common.dao.signup.SignupMapper;
+import com.beanions.common.dto.MembersDTO;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,5 +17,9 @@ public class SignupService {
 
     public int checkDupNkname(String nkname) {
         return signupMapper.checkDupNkname(nkname);
+    }
+
+    public int regist(MembersDTO member) {
+        return signupMapper.joinMember(member);
     }
 }
