+++ src/main/resources/templates/user/main.html
@@ -8,6 +8,12 @@
   <link rel="stylesheet" th:href="@{/css/user/main.css}">
   <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.js"></script>
   <title>YESINYERANG</title>
+  <script>
+    const message = `[[${ message }]]`;
+    if(message) {
+      alert(message);
+    }
+  </script>
 </head>
 <body>
 <div class="wrapper">
